Added TDD testing to get, getAll and delete services.Seperated testing.
