+++ auth-service/src/main/java/com/petclinic/auth/Exceptions/NotFoundException.java
@@ -1,0 +1,8 @@
+package com.petclinic.auth.Exceptions;
+
+public class NotFoundException  extends  RuntimeException{
+
+    public NotFoundException(String message) {
+        super(message);
+    }
+}
